+++ NEWS.md
@@ -1,5 +1,5 @@
 # arulesCBA 1.2.3 (2022-05-24)
-* install_LUCS_KDD_CMAR() and install_LUCS_KDD_CPAR(): Added more checks. This fixes timing issues with untar.
+* LUCS_KDD jars come now preinstalled
 
 # arulesCBA 1.2.1 (2021-11-20)
 * mineCARs now uses by default minimum LHS-support (via parameter originalSupport = FALSE).